refactor: 루트 및 파일, 컴포넌트 명 통일
