+++ client/src/Pages/CreateCrewingPage.jsx
@@ -1,0 +1,11 @@
+// eslint-disable-next-line no-unused-vars
+import React from 'react';
+import CreateCrewing from '../Components/CreateCrewing.jsx';
+
+export default function CreateCrewingPage() {
+  return (
+    <>
+      <CreateCrewing />
+    </>
+  );
+}
