+++ client/src/Pages/CreateDietPage.jsx
@@ -1,0 +1,11 @@
+// eslint-disable-next-line no-unused-vars
+import React from 'react';
+import CreateDiet from '../Components/CreateDiet.jsx';
+
+export default function CreateDietPage() {
+  return (
+    <>
+      <CreateDiet />
+    </>
+  );
+}
