+++ client/src/Pages/CreateWorkoutPage.jsx
@@ -1,0 +1,11 @@
+// eslint-disable-next-line no-unused-vars
+import React from 'react';
+import CreateWorkout from '../Components/CreateWorkout.jsx';
+
+export default function CreateWorkoutPage() {
+  return (
+    <>
+      <CreateWorkout />
+    </>
+  );
+}
