--- client/src/Pages/Makecrews.jsx
@@ -1,11 +1,0 @@
-// eslint-disable-next-line no-unused-vars
-import React from 'react';
-import Makecrew from '../Components/Makecrew.jsx';
-
-export default function Makeeats() {
-  return (
-    <>
-      <Makecrew />
-    </>
-  );
-}
