--- client/src/Pages/Makeeats.jsx
@@ -1,11 +1,0 @@
-// eslint-disable-next-line no-unused-vars
-import React from 'react';
-import Makeeat from '../Components/Makeeat.jsx';
-
-export default function Makeeats() {
-  return (
-    <>
-      <Makeeat />
-    </>
-  );
-}
