--- client/src/Pages/Makesports.jsx
@@ -1,11 +1,0 @@
-// eslint-disable-next-line no-unused-vars
-import React from 'react';
-import Makesport from '../Components/Makesport.jsx';
-
-export default function Makesports() {
-  return (
-    <>
-      <Makesport />
-    </>
-  );
-}
